requirements.txt
fastmcp>=0.1.0
PyYAML>=6.0
pytest>=8.0

